requirements.txt
feedparser==6.0.11
requests==2.31.0
praw==7.7.1
google-generativeai==0.7.2
python-dotenv==1.0.1
apscheduler==3.10.4
arxiv==2.1.0
